Carousel and Scrollspy
I made a first attempt at the carousel for the main page and a scrollspy for the purchases page. Neither worked quite as well as I had hoped, but they are at least there for us to work on.

I noticed that the font from the banner wasn't always rendering properly, so I changed the text box into a path in the .svg file, and that appears to have resolved the issue.

I copied the header and footer from the index page to all the other pages. For the time being, the header and the footer are working on the main page, but even though I did a simple copy-paste to the other pages, they refuse to work. I am not sure what the issue is. Perhaps it is a caching error on my end.

Finally I ought to mention that I added several photos to the repository to be used in the carousel.