Formatting Issues Resolved
I figured out the error with the pages other than index not accepting the header or the footer. It was due to an incorrect style sheet name.

I also dropped the idea of the scrollspy in favor of a big chart with a sticky navbar. It seems to work so far.