Simple Fixup
So I deleted all of the ~ files, I deleted the hammer picture, and I got the search page set up the way that it should have during the Zoom call.
I also put the image back into the join page the way it used to be. Hopefully this time we can get it to work.